chore(m3): drop -Wno-missing-field-initializers, quiet VK warning, docs
Review follow-ups:
- The Makefile suppression existed only because the VK GQA commit widened
  struct PC with act/alpha/limit push constants and left the existing
  positional initializers short. Name every field at all 16 sites (the
  gate_up dispatches overwrite act/alpha/limit right after, so semantics
  are unchanged -- the fields were implicitly zero before) and remove the
  flag from all five platform CFLAGS lines. gcc 15 builds warning-free.
- kv_alloc's index-cache calloc: (unsigned) cast gives value-range
  propagation a provable bound so -Walloc-size-larger-than stays quiet on
  inlined paths (the alloc is dev's own line; our wider "|| c->msa" guard
  changed inlining enough to surface it on the reviewer's gcc).
- docs/minimax-m3.md: family doc covering the shared-binary design, GQA/
  MSA/swigluoai/router notes, the CTX=4096 default and the MSA 2048-token
  exactness bound, tool-calling dialect, and the m3-tiny CI gate.

+++ c/colibri.c
@@ -7523,7 +7523,9 @@ static void kv_alloc(Model *m, int max_t){
         k->Lc8=k->Rc8=NULL; k->Lsc=k->Rsc=NULL; }
     if(k->Ic){ for(int i=0;i<c->n_layers;i++) free(k->Ic[i]); free(k->Ic); k->Ic=NULL; }
     if(m->has_dsa || c->msa){                        /* index-key cache: GLM DSA or M3 MSA */
-        k->Ic=calloc(c->n_layers,sizeof(float*));
+        /* (unsigned): n_layers is config-validated positive; the cast gives VRP a
+         * provable bound so -Walloc-size-larger-than stays quiet on inlined paths. */
+        k->Ic=calloc((unsigned)c->n_layers,sizeof(float*));
         for(int i=0;i<c->n_layers;i++) if(c->idx_type[i]) k->Ic[i]=falloc((int64_t)max_t*c->index_hd);
     }
     k->max_t=max_t;
